- cpu-only averaging for cuda backend - all-or-fail for p2p support
